Updated Packages

+++ Source Code/oratorz/lib/ui/pages/committee/widgets/dialogs/yield_spaeaker.dart
@@ -52,7 +52,7 @@ class _YieldSpeakerDialogState extends State<YieldSpeakerDialog> {
                       ),
                       onChanged: (value) => setState(() {
                         if (value != null) {
-                          selected = value as int;
+                          selected = value;
                         }
                       }),
                     ),
